Rewrite the jd-cmd code to use JD-Core 1.0.z API.
