+++ jd-lib/pom.xml
@@ -3,9 +3,9 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
-		<groupId>jd.core</groupId>
+		<groupId>com.github.kwart.jd</groupId>
 		<artifactId>jd-cmd</artifactId>
-		<version>0.9.3-SNAPSHOT</version>
+		<version>1.0.0-SNAPSHOT</version>
 	</parent>
 
 	<artifactId>jd-lib</artifactId>
@@ -13,55 +13,16 @@
 
 	<name>jd-lib</name>
 
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-	</properties>
-
-	<build>
-		<resources>
-			<resource>
-				<directory>src/main/resources</directory>
-				<targetPath>${project.build.outputDirectory}</targetPath>
-			</resource>
-			<resource>
-				<directory>src/main/java</directory>
-				<filtering>true</filtering>
-				<targetPath>../filtered-src</targetPath>
-			</resource>
-		</resources>
-		<sourceDirectory>target/filtered-src</sourceDirectory>
-
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.0</version>
-				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
 	<dependencies>
 		<dependency>
 			<groupId>org.slf4j</groupId>
 			<artifactId>slf4j-api</artifactId>
 			<version>${slf4j.version}</version>
+			<scope>provided</scope>
 		</dependency>
 		<dependency>
-			<groupId>jd</groupId>
+			<groupId>org.jd</groupId>
 			<artifactId>jd-core</artifactId>
-			<version>${jd.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>jd</groupId>
-			<artifactId>jd-common</artifactId>
-			<version>${jd.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>jd</groupId>
-			<artifactId>jd-common-ide</artifactId>
 			<version>${jd.version}</version>
 		</dependency>
 	</dependencies>
