+++ jd-lib/src/main/java/com/github/kwart/jd/loader/FileLoader.java
@@ -1,0 +1,27 @@
+package com.github.kwart.jd.loader;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+
+import org.jd.core.v1.api.loader.Loader;
+import org.jd.core.v1.api.loader.LoaderException;
+
+public class FileLoader implements Loader {
+
+    public static final FileLoader INSTANCE = new FileLoader();
+
+    @Override
+    public byte[] load(String internalName) throws LoaderException {
+        try {
+            return Files.readAllBytes(Paths.get(internalName));
+        } catch (IOException e) {
+            throw new LoaderException(e);
+        }
+    }
+
+    @Override
+    public boolean canLoad(String internalName) {
+        return Files.isReadable(Paths.get(internalName));
+    }
+}
