+++ pom.xml
@@ -2,15 +2,22 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>jd.core</groupId>
+	<groupId>com.github.kwart.jd</groupId>
 	<artifactId>jd-cmd</artifactId>
-	<version>0.9.3-SNAPSHOT</version>
+	<version>1.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 
 	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
+
+		<maven.shade.plugin.version>3.2.1</maven.shade.plugin.version>
+		<maven.clean.plugin.version>3.1.0</maven.clean.plugin.version>
 		<slf4j.version>1.7.5</slf4j.version>
-		<jd.version>0.7.1</jd.version>
+		<jd.version>1.0.7</jd.version>
 		<logback.version>1.0.13</logback.version>
+		<jcommander.version>1.32</jcommander.version>
 	</properties>
 
 	<modules>
@@ -22,7 +29,7 @@
 		<plugins>
 			<plugin>
 				<artifactId>maven-clean-plugin</artifactId>
-				<version>2.5</version>
+				<version>${maven.clean.plugin.version}</version>
 				<configuration>
 					<filesets>
 						<fileset>
@@ -41,8 +48,8 @@
 	
 	<repositories>
 		<repository>
-			<id>jd-core</id>
-			<url>https://raw.github.com/java-decompiler/mvn-repo/master/org/</url>
+			<id>JCenter</id>
+			<url>https://jcenter.bintray.com/</url>
 		</repository>
 	</repositories>
 </project>
